vk/innertia-soft-backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-decouple==3.8
pgvector==0.2.4
python-dateutil==2.8.2
Pillow==10.1.0
drf-spectacular==0.27.0
channels==4.0.0
channels-rest-framework==0.1.0
